Added TaskCommentController + Improved CommentServiceImpl

+++ src/main/java/com/example/Todo_list/service/impl/CommentServiceImpl.java
@@ -47,24 +47,40 @@ public class CommentServiceImpl implements CommentService {
 
     @Override
     public List<Comment> findCommentByUser(User user) {
+        if (user == null) {
+            logger.error("CommentService.findCommentByUser(): User cannot be null");
+            throw new NullEntityException(this.getClass().getName(), "User cannot be null");
+        }
+
         logger.info("CommentService.findCommentByUser(): Finding comments by " + user + " ...");
         return commentRepository.findByUser(user);
     }
 
     @Override
     public Comment updateComment(Comment comment) {
+        if (comment == null) {
+            logger.error("CommentService.updateComment(): Comment cannot be null");
+            throw new NullEntityException(this.getClass().getName(), "Comment cannot be null");
+        }
+
         logger.info("CommentService.updateComment(): Updating " + comment + " ...");
         return commentRepository.save(comment);
     }
 
     @Override
     public void deleteCommentById(Long id) {
+        Comment comment = this.findCommentById(id);
         logger.info("CommentService.deleteCommentById(): Deleting comment with id=" + id + " ...");
-        commentRepository.deleteById(id);
+        commentRepository.delete(comment);
     }
 
     @Override
     public void deleteComment(Comment comment) {
+        if (comment == null) {
+            logger.error("CommentService.deleteComment(): Comment cannot be null");
+            throw new NullEntityException(this.getClass().getName(), "Comment cannot be null");
+        }
+
         logger.info("CommentService.deleteComment(): Deleting " + comment + " ...");
         commentRepository.delete(comment);
     }
